# apps/daemon/pkg/toolbox/docs/swagger.yaml
definitions:
...
  CreateContextRequest:
    properties:
      cwd:
        type: string
      language:
        type: string
    type: object
...
  DaemonErrorCode:
    enum:
      - GIT_AUTH_FAILED
      - GIT_AUTH_FORBIDDEN
      - GIT_REPO_NOT_FOUND
      - GIT_BRANCH_NOT_FOUND
      - GIT_REF_NOT_FOUND
      - GIT_EMPTY_REPO
      - GIT_PUSH_REJECTED
      - GIT_DIRTY_WORKTREE
      - GIT_BRANCH_EXISTS
      - GIT_MERGE_CONFLICT
      - GIT_REPO_EXISTS
      - FILE_NOT_FOUND
      - FILE_ACCESS_DENIED
      - INVALID_FILE_PATH
      - LSP_SERVER_NOT_INITIALIZED
      - LSP_INVALID_REQUEST
      - PROCESS_EXECUTION_TIMEOUT
      - PROCESS_INVALID_COMMAND
      - PROCESS_NOT_FOUND
      - SESSION_ENDED
    type: string
    x-enum-varnames:
      - CodeGitAuthFailed
      - CodeGitAuthForbidden
      - CodeGitRepoNotFound
      - CodeGitBranchNotFound
      - CodeGitRefNotFound
      - CodeGitEmptyRepo
      - CodeGitPushRejected
      - CodeGitDirtyWorktree
      - CodeGitBranchExists
      - CodeGitMergeConflict
      - CodeGitRepoExists
      - CodeFileNotFound
      - CodeFileAccessDenied
      - CodeInvalidFilePath
      - CodeLspServerNotInitialized
      - CodeLspInvalidRequest
      - CodeProcessExecutionTimeout
      - CodeProcessInvalidCommand
      - CodeProcessNotFound
      - CodeSessionEnded
...
  ErrorResponse:
    description: Error response
    properties:
      code:
        allOf:
          - $ref: '#/definitions/DaemonErrorCode'
        example: GIT_REPO_NOT_FOUND
      message:
        example: Bad request
        type: string
      method:
        example: GET
        type: string
      path:
        example: /api/resource
        type: string
      source:
        example: DAYTONA_DAEMON
        type: string
      statusCode:
        example: 400
        type: integer
      timestamp:
        example: '2023-01-01T12:00:00Z'
        type: string
    required:
      - message
      - path
      - statusCode
      - timestamp
    type: object
...
  InterpreterContext:
    properties:
      active:
        type: boolean
      createdAt:
        type: string
      cwd:
        type: string
      id:
        type: string
      language:
        type: string
    required:
      - active
      - createdAt
      - cwd
      - id
      - language
    type: object
...
  ListContextsResponse:
    properties:
      contexts:
        items:
          $ref: '#/definitions/InterpreterContext'
        type: array
    required:
      - contexts
    type: object
...
  PtyCreateRequest:
    properties:
      cols:
        type: integer
      cwd:
        type: string
      envs:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      lazyStart:
        description: Don't start PTY until first client connects
        type: boolean
      rows:
        type: integer
    type: object
  PtyCreateResponse:
    properties:
      sessionId:
        type: string
    required:
      - sessionId
    type: object
  PtyListResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/PtySessionInfo'
        type: array
    required:
      - sessions
    type: object
  PtyResizeRequest:
    properties:
      cols:
        maximum: 1000
        minimum: 1
        type: integer
      rows:
        maximum: 1000
        minimum: 1
        type: integer
    required:
      - cols
      - rows
    type: object
  PtySessionInfo:
    properties:
      active:
        type: boolean
      cols:
        type: integer
      createdAt:
        type: string
      cwd:
        type: string
      envs:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      lazyStart:
        description: Whether this session uses lazy start
        type: boolean
      rows:
        type: integer
    required:
      - active
      - cols
      - createdAt
      - cwd
      - envs
      - id
      - lazyStart
      - rows
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
info:
  contact: {}
  description: Daytona Toolbox API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Daytona Toolbox API
  version: v0.0.0-dev
paths:
...
  /process/interpreter/context:
    get:
      description: Returns information about all user-created interpreter contexts
        (excludes default context)
      operationId: ListInterpreterContexts
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ListContextsResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List all user-created interpreter contexts
      tags:
        - interpreter
    post:
      consumes:
        - application/json
      description: Creates a new isolated interpreter context with optional working
        directory and language
      operationId: CreateInterpreterContext
      parameters:
        - description: Context configuration
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/CreateContextRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/InterpreterContext'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a new interpreter context
      tags:
        - interpreter
  /process/interpreter/context/{id}:
    delete:
      description: Deletes an interpreter context and shuts down its worker process
      operationId: DeleteInterpreterContext
      parameters:
        - description: Context ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete an interpreter context
      tags:
        - interpreter
...
  /process/pty:
    get:
      description: Get a list of all active pseudo-terminal sessions
      operationId: ListPtySessions
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PtyListResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List all PTY sessions
      tags:
        - process
    post:
      consumes:
        - application/json
      description: Create a new pseudo-terminal session with specified configuration
      operationId: CreatePtySession
      parameters:
        - description: PTY session creation request
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/PtyCreateRequest'
      produces:
        - application/json
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/PtyCreateResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a new PTY session
      tags:
        - process
  /process/pty/{sessionId}:
    delete:
      description: Delete a pseudo-terminal session and terminate its process
      operationId: DeletePtySession
      parameters:
        - description: PTY session ID
          in: path
          name: sessionId
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete a PTY session
      tags:
        - process
    get:
      description: Get detailed information about a specific pseudo-terminal session
      operationId: GetPtySession
      parameters:
        - description: PTY session ID
          in: path
          name: sessionId
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PtySessionInfo'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get PTY session information
      tags:
        - process
  /process/pty/{sessionId}/connect:
    get:
      description: Establish a WebSocket connection to interact with a pseudo-terminal
        session
      operationId: ConnectPtySession
      parameters:
        - description: PTY session ID
          in: path
          name: sessionId
          required: true
          type: string
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Connect to PTY session via WebSocket
      tags:
        - process
  /process/pty/{sessionId}/resize:
    post:
      consumes:
        - application/json
      description: Resize the terminal dimensions of a pseudo-terminal session
      operationId: ResizePtySession
      parameters:
        - description: PTY session ID
          in: path
          name: sessionId
          required: true
          type: string
        - description: Resize request with new dimensions
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/PtyResizeRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PtySessionInfo'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '410':
          description: Gone
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Resize a PTY session
      tags:
        - process
...
swagger: '2.0'