# libs/toolbox-api-client-ruby/lib/daytona_toolbox_api_client/models/daemon_error_code.rb
=begin
#Daytona Toolbox API

#Daytona Toolbox API

The version of the OpenAPI document: v0.0.0-dev

Generated by: https://openapi-generator.tech
Generator version: 7.21.0

=end

require 'date'
require 'time'

module DaytonaToolboxApiClient
  class DaemonErrorCode
    CodeGitAuthFailed = "GIT_AUTH_FAILED".freeze
    CodeGitAuthForbidden = "GIT_AUTH_FORBIDDEN".freeze
    CodeGitRepoNotFound = "GIT_REPO_NOT_FOUND".freeze
    CodeGitBranchNotFound = "GIT_BRANCH_NOT_FOUND".freeze
    CodeGitRefNotFound = "GIT_REF_NOT_FOUND".freeze
    CodeGitEmptyRepo = "GIT_EMPTY_REPO".freeze
    CodeGitPushRejected = "GIT_PUSH_REJECTED".freeze
    CodeGitDirtyWorktree = "GIT_DIRTY_WORKTREE".freeze
    CodeGitBranchExists = "GIT_BRANCH_EXISTS".freeze
    CodeGitMergeConflict = "GIT_MERGE_CONFLICT".freeze
    CodeGitRepoExists = "GIT_REPO_EXISTS".freeze
    CodeFileNotFound = "FILE_NOT_FOUND".freeze
    CodeFileAccessDenied = "FILE_ACCESS_DENIED".freeze
    CodeInvalidFilePath = "INVALID_FILE_PATH".freeze
    CodeLspServerNotInitialized = "LSP_SERVER_NOT_INITIALIZED".freeze
    CodeLspInvalidRequest = "LSP_INVALID_REQUEST".freeze
    CodeProcessExecutionTimeout = "PROCESS_EXECUTION_TIMEOUT".freeze
    CodeProcessInvalidCommand = "PROCESS_INVALID_COMMAND".freeze
    CodeProcessNotFound = "PROCESS_NOT_FOUND".freeze
    CodeSessionEnded = "SESSION_ENDED".freeze
    UNKNOWN_DEFAULT_OPEN_API = "unknown_default_open_api".freeze

    def self.all_vars
      @all_vars ||= [CodeGitAuthFailed, CodeGitAuthForbidden, CodeGitRepoNotFound, CodeGitBranchNotFound, CodeGitRefNotFound, CodeGitEmptyRepo, CodeGitPushRejected, CodeGitDirtyWorktree, CodeGitBranchExists, CodeGitMergeConflict, CodeGitRepoExists, CodeFileNotFound, CodeFileAccessDenied, CodeInvalidFilePath, CodeLspServerNotInitialized, CodeLspInvalidRequest, CodeProcessExecutionTimeout, CodeProcessInvalidCommand, CodeProcessNotFound, CodeSessionEnded, UNKNOWN_DEFAULT_OPEN_API].freeze
    end

    # Builds the enum from string
    # @param [String] The enum value in the form of the string
    # @return [String] The enum value
    def self.build_from_hash(value)
      new.build_from_hash(value)
    end

    # Builds the enum from string
    # @param [String] The enum value in the form of the string
    # @return [String] The enum value
    def build_from_hash(value)
      return value if DaemonErrorCode.all_vars.include?(value)
      UNKNOWN_DEFAULT_OPEN_API
    end
  end
end