# libs/toolbox-api-client/src/models/daemon-error-code.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Daytona Toolbox API
 * Daytona Toolbox API
 *
 * The version of the OpenAPI document: v0.0.0-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const DaemonErrorCode = {
    CodeGitAuthFailed: 'GIT_AUTH_FAILED',
    CodeGitAuthForbidden: 'GIT_AUTH_FORBIDDEN',
    CodeGitRepoNotFound: 'GIT_REPO_NOT_FOUND',
    CodeGitBranchNotFound: 'GIT_BRANCH_NOT_FOUND',
    CodeGitRefNotFound: 'GIT_REF_NOT_FOUND',
    CodeGitEmptyRepo: 'GIT_EMPTY_REPO',
    CodeGitPushRejected: 'GIT_PUSH_REJECTED',
    CodeGitDirtyWorktree: 'GIT_DIRTY_WORKTREE',
    CodeGitBranchExists: 'GIT_BRANCH_EXISTS',
    CodeGitMergeConflict: 'GIT_MERGE_CONFLICT',
    CodeGitRepoExists: 'GIT_REPO_EXISTS',
    CodeFileNotFound: 'FILE_NOT_FOUND',
    CodeFileAccessDenied: 'FILE_ACCESS_DENIED',
    CodeInvalidFilePath: 'INVALID_FILE_PATH',
    CodeLspServerNotInitialized: 'LSP_SERVER_NOT_INITIALIZED',
    CodeLspInvalidRequest: 'LSP_INVALID_REQUEST',
    CodeProcessExecutionTimeout: 'PROCESS_EXECUTION_TIMEOUT',
    CodeProcessInvalidCommand: 'PROCESS_INVALID_COMMAND',
    CodeProcessNotFound: 'PROCESS_NOT_FOUND',
    CodeSessionEnded: 'SESSION_ENDED',
    UNKNOWN_DEFAULT_OPEN_API: '11184809',
} as const;

export type DaemonErrorCode = typeof DaemonErrorCode[keyof typeof DaemonErrorCode];